# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления бронированиями",
        "title": "Booker Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/bookingsEntity/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить детальную информацию о объекте бронирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsEntity"
                ],
                "summary": "Получить объект бронирования по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объекта бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_entities_get_booking_entity.BookingEntityResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновить объект бронирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsEntity"
                ],
                "summary": "Обновить объект бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объекта бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные типа бронирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_entities_create_booking_entity.BookingEntity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить объект бронирования по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsEntity"
                ],
                "summary": "Удалить объект бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объекта бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/bookingsType": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список всех типов бронирований с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsType"
                ],
                "summary": "Получить список типов бронирований",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Сортировка по id. asc, desc",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по name. asc, desc",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка по description. asc, desc",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Лимит на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_list.BookingTypeList"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создать тип бронирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsType"
                ],
                "summary": "Создать тип бронирования",
                "parameters": [
                    {
                        "description": "Данные типа бронирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.CreateBookingTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId"
                        }
                    }
                }
            }
        },
...
        "/bookingsType/{id}/bookingsEntity": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список всех объектов бронирований с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsEntity"
                ],
                "summary": "Получить список объектов бронирований у определённого типа бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Лимит на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_entities_get_booking_type_entities.BookingTypeEntitiesResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_entities_create_booking_entity.BookingEntity": {
            "type": "object",
            "properties": {
                "booking_type_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_entities_get_booking_entity.BookingEntityResponse": {
            "type": "object",
            "properties": {
                "booking_type_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking_entities_get_booking_type_entities.BookingTypeEntitiesResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.CreateBookingTypeRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_list.BookingTypeInfoList": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_list.BookingTypeList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_list.BookingTypeInfoList"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_list.BookingTypeListMetaData"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_list.BookingTypeListMetaData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add \"Bearer\" before token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}